MYR-77 Split audit sidecar wiring into wiring_audit.go (300-line cap)
Extract buildAuditRepo + setupAuditSidecar from wiring.go into
wiring_audit.go so wiring.go stays at exactly 300 lines per CLAUDE.md cap.
No functional change.
